UPBGE: Restore object info node.
Previously the object info node was doing nothing directly by the fact
that GPU_material_bind_uniforms was using nullptr for the last argument
corresponding to the float array of object info.

The pass indices used by the object info is stored either in the
blender material and the game object under the same variable m_passIndex.
The object is coppying this value to the mesh user and BL_BlenderShader::Update
is called with the material pass index as argument and inside it
get the object pass index from the mesh user. These information are
used to set the object info and the random value is computed from
the hash of the mesh user client object, a per object value.

The pass index is exposed in python too under attributes:
KX_BlenderMaterial.passIndex
KX_GameObject.passIndex